Factored out `TransitionSystemEncoding` and SMT helpers in `bmc.rs` into separate files; factored `expr_at_step` into new `step_at` function in `TransitionSystemEncoding` trait
