+++ patronus/src/mc.rs
@@ -3,12 +3,12 @@
 // author: Kevin Laeufer <laeufer@berkeley.edu>
 
 mod bmc;
+mod encoding;
 mod pdr;
 mod types;
+mod utils;
 
-pub use bmc::{
-    ModelCheckResult, TransitionSystemEncoding, UnrollSmtEncoding, bmc, check_assuming,
-    check_assuming_end, get_smt_value,
-};
+pub use bmc::bmc;
 pub use pdr::pdr;
-pub use types::{InitValue, Witness};
+pub use types::{InitValue, ModelCheckResult, Witness};
+pub use utils::get_smt_value;
