+++ patronus/src/mc/pdr.rs
@@ -4,18 +4,16 @@
 
 use crate::expr::*;
 use crate::mc::bmc::start_bmc_or_pdr;
-use crate::mc::{
-    ModelCheckResult, TransitionSystemEncoding, bmc, check_assuming, check_assuming_end,
-    get_smt_value,
-};
+use crate::mc::encoding::{Step, TransitionSystemEncoding};
+use crate::mc::types::Result;
+use crate::mc::utils::{check_assuming, check_assuming_end, get_smt_value};
+use crate::mc::{ModelCheckResult, bmc};
 use crate::smt::*;
 use crate::system::TransitionSystem;
 use baa::{BitVecOps, Value};
 use std::collections::BinaryHeap;
 use std::num::NonZeroUsize;
 use std::ops::{Index, IndexMut};
-
-type Step = u64;
 
 const FROM_STEP: Step = 1;
 
@@ -164,25 +162,6 @@ impl FrameId {
 // PDR helper functions
 // -------------------------------------------------------------------------------------------------
 
-/// Get the stepped version of an SMT expression
-///
-/// # Preconditions
-/// * `expr` must exist in `enc` at `step`
-fn expr_at_step(
-    ctx: &mut Context,
-    enc: &impl TransitionSystemEncoding,
-    expr: ExprRef,
-    step: Step,
-) -> ExprRef {
-    simple_transform_expr(ctx, expr, |ctx, e, _| {
-        if ctx[e].is_symbol() {
-            Some(enc.get_at(ctx, e, step))
-        } else {
-            None
-        }
-    })
-}
-
 /// Extract states values from solver at a certain time step
 ///
 /// # Preconditions
@@ -202,7 +181,7 @@ fn extract_state_values(
 
     // Extract exact SMT value for each system state
     for state in &sys.states {
-        let sym = enc.get_at(ctx, state.symbol, step);
+        let sym = enc.step_at(ctx, state.symbol, step);
         state_vals.push((state.symbol, get_smt_value(ctx, smt_ctx, sym)?));
     }
 
@@ -306,7 +285,7 @@ impl Frame {
     ) -> Result<()> {
         // Create implication act_i => \neg c, where c is the blocked cube stepped at current step
         let clause = cube.negate(ctx);
-        let clause_expr = expr_at_step(ctx, enc, clause, FROM_STEP);
+        let clause_expr = enc.step_at(ctx, clause, FROM_STEP);
         let imp = ctx.implies(self.act, clause_expr);
 
         // Permanently assert implication in solver
@@ -400,7 +379,7 @@ impl BasePdr {
 
         // Always assert constraints at current step
         for &cons in &sys.constraints {
-            let cons_cur = expr_at_step(ctx, enc, cons, FROM_STEP);
+            let cons_cur = enc.step_at(ctx, cons, FROM_STEP);
             smt_ctx.assert(ctx, cons_cur)?;
         }
 
@@ -410,7 +389,7 @@ impl BasePdr {
         let cons_next_expr = sys
             .constraints
             .iter()
-            .map(|&c| expr_at_step(ctx, enc, c, TO_STEP))
+            .map(|&c| enc.step_at(ctx, c, TO_STEP))
             .collect::<Vec<_>>()
             .into_iter()
             .fold(ctx.get_true(), |acc, c| ctx.and(acc, c));
@@ -425,7 +404,7 @@ impl BasePdr {
 
         // Create act_0 => init_cube, where init_cube is stepped to the before step
         let init_expr = init_cube.to_expr(ctx);
-        let init_expr = expr_at_step(ctx, enc, init_expr, FROM_STEP);
+        let init_expr = enc.step_at(ctx, init_expr, FROM_STEP);
         let init_imp = ctx.implies(init_act, init_expr);
 
         // Permanently assert implication in solver
@@ -519,13 +498,13 @@ impl BasePdr {
 
         // Next step cube
         let cube_expr = cube.cube.to_expr(ctx);
-        let cube_next = expr_at_step(ctx, enc, cube_expr, TO_STEP);
+        let cube_next = enc.step_at(ctx, cube_expr, TO_STEP);
         assumptions.push(cube_next);
 
         // Current step negation cube
         if query_type == RelIndType::Extended {
             let neg_cube_expr = cube.cube.negate(ctx);
-            let neg_cube_cur = expr_at_step(ctx, enc, neg_cube_expr, FROM_STEP);
+            let neg_cube_cur = enc.step_at(ctx, neg_cube_expr, FROM_STEP);
             assumptions.push(neg_cube_cur);
         }
 
@@ -567,7 +546,7 @@ impl BasePdr {
         let bad_lits: Vec<ExprRef> = sys
             .bad_states
             .iter()
-            .map(|&b| expr_at_step(ctx, enc, b, FROM_STEP))
+            .map(|&b| enc.step_at(ctx, b, FROM_STEP))
             .collect();
 
         // Disjunct all bad state literals
@@ -754,10 +733,10 @@ impl BasePdr {
             let inf_assumption = self.frame_assumptions(ctx, FrameId::Infinite);
 
             let clause_expr = cube.negate(ctx);
-            let clause_cur = expr_at_step(ctx, enc, clause_expr, FROM_STEP);
+            let clause_cur = enc.step_at(ctx, clause_expr, FROM_STEP);
 
             let cube_expr = cube.to_expr(ctx);
-            let cube_next = expr_at_step(ctx, enc, cube_expr, TO_STEP);
+            let cube_next = enc.step_at(ctx, cube_expr, TO_STEP);
 
             // Run the query `SAT?[R_\infty /\ \neg c /\ T /\ c']`, asserting that the
             // constraints hold at `TO_STEP`
